generate interface doc ui fixes and refactoring
