+++ frontend/src/components/features/home/tools/tool-modal.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import {
   Dialog,
   DialogTitle as RawDialogTitle,
@@ -6,18 +6,13 @@ import {
 } from "#/components/ui/dialog";
 import styles from "./ToolCard.module.css";
 import { BrandButton } from "../../settings/brand-button";
-import { form, VisuallyHidden } from "@heroui/react";
+import { VisuallyHidden } from "@heroui/react";
 import { SettingsInput } from "../../settings/settings-input";
-import { useUserProviders } from "#/hooks/use-user-providers";
 import { useCreateConversation } from "#/hooks/mutation/use-create-conversation";
-import { useTranslation } from "react-i18next";
 import { useWorkspace } from "#/context/WorkspaceContext";
 import { RepositorySelectionForm } from "../repo-selection-form";
 import { dataSourceToGitRepository } from "#/utils/utils.ts";
-import { SettingsDropdownInput } from "../../settings/settings-dropdown-input";
-import { generateRepoDocumentationForAWorkspace } from "#/api/workspaces";
-import { toast } from "sonner";
-import { getAllDataSourcesByWorkspaceId } from "#/api/data-sources";
+import GenerateInterfaceDocForm from "./generate-interface-documentation-form";
 
 const DialogContent = RawDialogContent as React.FC<
   React.PropsWithChildren<any>
@@ -183,98 +178,5 @@ export function ToolModal({
         </div>
       </DialogContent>
     </Dialog>
-  );
-}
-
-function GenerateInterfaceDocForm() {
-  const { workspaces } = useWorkspace();
-
-  const workspaceItems = workspaces.map((workspace) => ({
-    key: workspace.id,
-    label: workspace.name,
-  }));
-
-  const [selectedWorkspaceKey, setSelectedWorkspaceKey] = useState<
-    string | undefined
-  >(undefined);
-  const [lastProcessedRepoVersion, setLastProcessedRepoVersion] = useState<
-    Record<string, any> | undefined
-  >(undefined);
-
-  const generateRepoDocumentation = async () => {
-    if (!selectedWorkspaceKey) {
-      return;
-    }
-
-    const { data, success, errorMessage } =
-      await generateRepoDocumentationForAWorkspace(selectedWorkspaceKey);
-    if (success) {
-      toast.success(data.message || "Documentation generation started");
-    } else {
-      toast.error(errorMessage ?? "Documentation generation failed");
-    }
-  };
-
-  function getLastProcessedVersion(repo: any) {
-    const processedVersions =
-      repo?.versions?.filter((v: any) => v.status === "PROCESSED") || [];
-    return processedVersions.length ? processedVersions.at(-1) : null;
-  }
-
-  const getAllDataSourcesForWorkspace = async () => {
-    const { data, success, errorMessage } =
-      await getAllDataSourcesByWorkspaceId(selectedWorkspaceKey);
-    if (success) {
-      const repositories = data?.filter(
-        (source: any) => source.type === "GIT_REPOSITORY",
-      );
-      const lastProcessedVersion = getLastProcessedVersion(repositories?.[0]);
-      setLastProcessedRepoVersion(lastProcessedVersion);
-    } else {
-      toast.error(errorMessage);
-    }
-  };
-
-  useEffect(() => {
-    getAllDataSourcesForWorkspace();
-  }, [selectedWorkspaceKey]);
-
-  return (
-    <div className="w-full flex flex-col items-center">
-      <SettingsDropdownInput
-        testId="workspace-interface-dropdown"
-        name="workspace-interface-dropdown"
-        placeholder="Select a workspace"
-        items={workspaceItems}
-        isDisabled={false}
-        wrapperClassName="w-full"
-        selectedKey={selectedWorkspaceKey}
-        onSelectionChange={(key) => {
-          setSelectedWorkspaceKey(key?.toString());
-        }}
-      />
-
-      {selectedWorkspaceKey &&
-        (lastProcessedRepoVersion ? (
-          <p className="my-4">
-            Last processed Repo Version: ${lastProcessedRepoVersion.name}
-          </p>
-        ) : (
-          <p className="my-4">No repo version found.</p>
-        ))}
-
-      <div className="flex justify-center w-full">
-        <BrandButton
-          testId="tool-generate-button"
-          variant="primary"
-          type="button"
-          className="mt-4 max-w-md w-full text-lg font-bold"
-          isDisabled={!selectedWorkspaceKey || !lastProcessedRepoVersion}
-          onClick={generateRepoDocumentation}
-        >
-          {false ? "Generating.." : "Generate"}
-        </BrandButton>
-      </div>
-    </div>
   );
 }
